refactor: 스타일 유틸리티 파일 구조 및 내보내기 정리
