+++ src/lib/index.ts
@@ -1,8 +1,3 @@
-// constants
 export * from './constants';
-
-// utils
-export * from './utils';
-
-// types
 export * from './types';
+export * from './utils';
