+++ src/lib/utils/style/index.ts
@@ -1,1 +1,1 @@
-export * from './css';
+export * from './jsx-style-to-css';
